Try to make PHP-FPM (the wiki) more robust
Apparently, it's a good idea to use pm.max_requests to make workers restart every so often because PHP is so useless.  Hopefully, this will prevent the random shutdown we experienced this morning.

I've also added a public /ping page and a private /status page to the OAuth wiki, allowing me to quickly and easily check the status of the PHP-FPM container itself.